Move ChildDrawingOrderStrategy interface to separate file
